{{cookiecutter.project_slug}}/requirements.txt
Django==3.2.12
django-globus-portal-framework==0.4.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
gunicorn==20.1.0
humanize==4.0.0
ssm-parameter-store==19.11.0
whitenoise==6.0.0